Move struct `file_port` definition into new header `file_port.hpp`
