+++ src/kernel/library.cpp
@@ -17,12 +17,11 @@
 #include <chrono>
 #include <filesystem>
 #include <numeric>
-#include <string_view>
 
 #include <meevax/kernel/basis.hpp>
 #include <meevax/kernel/disassemble.hpp>
+#include <meevax/kernel/file_port.hpp>
 #include <meevax/kernel/library.hpp>
-#include <meevax/utility/debug.hpp>
 
 namespace meevax
 {
